fix(compiler): ambiguous conflict updates, unguarded membership, count typing
Five bugs found porting api-v5, four of which compiled cleanly and only failed
against a real database. Each fix below was checked against Postgres 17.

Inside `ON CONFLICT ... DO UPDATE` two rows are in scope — the existing one under
the table's name and the proposed one under EXCLUDED — so a bare column on the
right-hand side is ambiguous and Postgres rejects the statement with 42702. Even
the degenerate `else (update User set { email := .email })` failed. Field
references there are now qualified with the target table; the assignment target
stays bare, which is what Postgres requires.

`$x? in .someMultiLink` returned early from the membership branch and so skipped
the optional-param null guard every other position gets. An omitted param made
`EXISTS` over `IN (NULL)` false, so the filter matched *nothing* where
aql/parameters/optional.md promises match-all. The guard is extracted and the
membership result routed through it.

`$x in .link.multiScalar` compiles its right-hand side to a correlated scalar
subquery, and `ANY (<subquery>)` is the subquery form of ANY: Postgres reads it
as a set of rows and rejects `text = text[]`. Casting the operand makes it an
ordinary array expression and selects the array form.

A correlated aggregate was accepted in `order by` but rejected as a computed
shape field with `unknown type ""` — the aggregate carries its type in AggFunc,
not TypeName, and the shape path resolved TypeName regardless. It now dispatches
to the same compileAggSQL the order-by position already used.

COUNT is bigint, which every driver hands back as a string to preserve precision
past 2^53, so a generated Promise<number> was a lie and a paginated `total`
silently became "2". Cast to INTEGER and keep the typing honest.

Also: casts are postfix, but the parser reports a prefix cast at the *next*
token, several characters past the mistake. Parse errors now carry a hint naming
the construct.

TestInOnMultiScalarThroughSingleLink asserted the un-cast form on the assumption
that "ANY accepts a correlated subquery". It does not; the test was codifying the
bug and now asserts the cast.

+++ core/compiler/compiler_test.go
@@ -178,7 +178,9 @@ type Place {
 	if err != nil {
 		t.Fatalf("compile subquery aggregate query: %v", err)
 	}
-	if !strings.Contains(compiledSub.SQL, `(SELECT COUNT(*) FROM (`) {
+	// COUNT is bigint, which the drivers hand back as a string; the cast keeps the
+	// generated Promise<number>/int64 typing honest.
+	if !strings.Contains(compiledSub.SQL, `(SELECT COUNT(*)::INTEGER FROM (`) {
 		t.Errorf("expected subquery count compilation, got:\n%s", compiledSub.SQL)
 	}
